+++ sql/update-fuseki.sql
@@ -1,0 +1,105 @@
+\pset null '(null)'
+
+CREATE SERVER fuseki
+FOREIGN DATA WRAPPER rdf_fdw 
+OPTIONS (
+  endpoint   'http://fuseki:3030/dt/sparql',
+  update_url 'http://fuseki:3030/dt/update');
+
+CREATE FOREIGN TABLE ft (
+  subject   rdfnode OPTIONS (variable '?s'),
+  predicate rdfnode OPTIONS (variable '?p'),
+  object    rdfnode OPTIONS (variable '?o') 
+)
+SERVER fuseki OPTIONS (
+  log_sparql 'true',
+  sparql 'SELECT * WHERE {?s ?p ?o}',
+  sparql_update_pattern '?s ?p ?o .'
+);
+
+CREATE USER MAPPING FOR postgres
+SERVER fuseki OPTIONS (user 'admin', password 'secret');
+
+INSERT INTO ft (subject, predicate, object)
+VALUES  ('<https://www.uni-muenster.de>', '<http://www.w3.org/2000/01/rdf-schema#label>', '"Westfälische Wilhelms-Universität Münster"@de');
+
+SELECT * FROM ft;
+
+UPDATE ft SET
+  object = '"University of Münster"@en'
+WHERE subject = '<https://www.uni-muenster.de>'
+  AND predicate = '<http://www.w3.org/2000/01/rdf-schema#label>';
+
+SELECT * FROM ft;
+
+UPDATE ft 
+SET object = '""'::rdfnode
+WHERE subject = '<https://www.uni-muenster.de>'
+  AND predicate = '<http://www.w3.org/2000/01/rdf-schema#label>';
+SELECT * FROM ft;
+
+UPDATE ft 
+SET object = '🐘'::rdfnode
+WHERE subject = '<https://www.uni-muenster.de>'
+  AND predicate = '<http://www.w3.org/2000/01/rdf-schema#label>';
+SELECT * FROM ft;
+
+/* update using existing column */
+UPDATE ft 
+SET object = predicate
+WHERE subject = '<https://www.uni-muenster.de>'
+  AND predicate = '<http://www.w3.org/2000/01/rdf-schema#label>';
+SELECT * FROM ft;
+
+/* update literals with quotes */
+UPDATE ft 
+SET object = '"\"text with quotes\""'::rdfnode
+WHERE subject = '<https://www.uni-muenster.de>'
+  AND predicate = '<http://www.w3.org/2000/01/rdf-schema#label>';
+SELECT * FROM ft;
+
+/* update literals containing newlines */
+UPDATE ft 
+SET object = '"text \n newline"'::rdfnode
+WHERE subject = '<https://www.uni-muenster.de>'
+  AND predicate = '<http://www.w3.org/2000/01/rdf-schema#label>';
+SELECT * FROM ft;
+
+/* update to xsd:string literal (Fuseki might omit xsd:string datatype) */
+UPDATE ft 
+SET object = '"text xsd string"^^<http://www.w3.org/2001/XMLSchema#string>'::rdfnode
+WHERE subject = '<https://www.uni-muenster.de>'
+  AND predicate = '<http://www.w3.org/2000/01/rdf-schema#label>';
+SELECT * FROM ft;
+
+/* update to xsd:int literal */
+UPDATE ft 
+SET object = 42::rdfnode
+WHERE subject = '<https://www.uni-muenster.de>'
+  AND predicate = '<http://www.w3.org/2000/01/rdf-schema#label>';
+SELECT * FROM ft;
+
+/* update to xsd:decimal literal */
+UPDATE ft 
+SET object = 42.37::rdfnode
+WHERE subject = '<https://www.uni-muenster.de>'
+  AND predicate = '<http://www.w3.org/2000/01/rdf-schema#label>';
+SELECT * FROM ft;
+
+/* update to xsd:long literal */
+UPDATE ft 
+SET object = 423712345678911::rdfnode
+WHERE subject = '<https://www.uni-muenster.de>'
+  AND predicate = '<http://www.w3.org/2000/01/rdf-schema#label>';
+SELECT * FROM ft;
+
+/* update with NULL - must fail (after SELECT since FDW must fetch OLD values first) */
+UPDATE ft 
+SET object = NULL
+WHERE subject = '<https://www.uni-muenster.de>'
+  AND predicate = '<http://www.w3.org/2000/01/rdf-schema#label>';
+SELECT * FROM ft;
+
+/* cleanup */
+DELETE FROM ft;
+DROP SERVER fuseki CASCADE;